Add a few missing constraints to PlanConstraints. Change visibility of plan and plan constraints.
