+++ src/plan/gencopy/gc_works.rs
@@ -1,5 +1,5 @@
 use super::global::GenCopy;
-use crate::plan::global::PlanConstraints;
+use crate::plan::PlanConstraints;
 use crate::plan::CopyContext;
 use crate::policy::space::Space;
 use crate::scheduler::gc_works::*;
