+++ src/plan/gencopy/mod.rs
@@ -3,3 +3,4 @@ pub(super) mod global;
 pub(super) mod mutator;
 
 pub use self::global::GenCopy;
+pub use self::global::GENCOPY_CONSTRAINTS;
