+++ src/plan/nogc/mod.rs
@@ -2,3 +2,4 @@ pub(super) mod global;
 pub(super) mod mutator;
 
 pub use self::global::NoGC;
+pub use self::global::NOGC_CONSTRAINTS;
