+++ src/plan/semispace/mod.rs
@@ -3,3 +3,4 @@ pub(super) mod global;
 pub(super) mod mutator;
 
 pub use self::global::SemiSpace;
+pub use self::global::SS_CONSTRAINTS;
